database_migrate/requirements.txt
pandas==2.1.4
pyodbc==5.0.1
pymssql==2.2.11
configparser==6.0.0
logging
tqdm==4.66.1
click==8.1.7
pyyaml==6.0.1 